mv isServerSideIssue() and logServerError() to result helpers.
